Flatten base router, rename
