+++ core/server/services/route/index.js
@@ -20,4 +20,3 @@
 module.exports.appRouter = require('./app-router');
 // This is the main router, that gets mounted in the express app in /site
 module.exports.router = require('./site-router');
-
